Change lround raise_invalid* function type from int to long

+++ usr/src/libm/src/m9x/lround.c
@@ -83,7 +83,7 @@ lround(double x) {
 static const double dtype_min = LONG_MIN - 0.5;
 static const double dtype_max = LONG_MAX + 0.5;
 
-static int
+static long
 raise_invalid_with_max_long() {
 	static const double zero = 0.0;
 	volatile double dummy  __attribute__((unused));
